Fix type of PuglEventText::time

+++ pugl/pugl.h
@@ -308,7 +308,7 @@ typedef struct {
 typedef struct {
 	PuglEventType type;        /**< PUGL_CHAR. */
 	uint32_t      flags;       /**< Bitwise OR of PuglEventFlag values. */
-	uint32_t      time;        /**< Time in milliseconds. */
+	double        time;        /**< Time in milliseconds. */
 	double        x;           /**< View-relative X coordinate. */
 	double        y;           /**< View-relative Y coordinate. */
 	double        x_root;      /**< Root-relative X coordinate. */
